# dashboard/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: MONOLITH OS API
  description: |
    RESTful API for MONOLITH OS - An AI-powered enterprise operating system that orchestrates
    autonomous agents for business operations, decision-making, and workflow management.

    ## Features
    - **Task Management**: Create, update, and track tasks across the organization
    - **Decision Logging**: Record and audit business decisions with full traceability
    - **Workflow Orchestration**: Manage multi-step business processes
    - **Agent Management**: Monitor and control AI agents across different roles

    ## Authentication
    All endpoints require Bearer token authentication. Obtain a token via `/auth/login`.

    ## Rate Limiting
    - Standard endpoints: 100 requests/minute
    - Bulk operations: 10 requests/minute

    ## Versioning
    The API uses URL-based versioning. Current version: v1
  version: 1.0.0
  contact:
    name: MONOLITH OS Support
    email: support@monolith-os.local
  license:
    name: Proprietary
    url: https://monolith-os.local/license
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.monolith-os.local/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /v1/workflows:
    get:
      tags:
        - Workflows
      summary: List workflows
      description: Retrieve a paginated list of workflows
      parameters:
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [active, paused, completed, cancelled, draft]
        - name: type
          in: query
          description: Filter by workflow type
          schema:
            type: string
            enum: [operational, strategic, compliance, hr, financial, technical]
        - name: owner_role
          in: query
          description: Filter by owner role
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Workflows
      summary: Create a workflow
      description: Create a new workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowCreate'
      responses:
        '201':
          description: Workflow created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  parameters:
...
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Workflow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, paused, completed, cancelled, draft]
        type:
          type: string
          enum: [operational, strategic, compliance, hr, financial, technical]
        owner_role:
          type: string
        owner_name:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        start_date:
          type: string
          format: date-time
          nullable: true
        target_date:
          type: string
          format: date-time
          nullable: true
        tasks_total:
          type: integer
        tasks_completed:
          type: integer
        created_at:
          type: string
          format: date-time
    WorkflowCreate:
      type: object
      required:
        - name
        - owner_role
      properties:
        name:
          type: string
          description: Workflow name
        description:
          type: string
        type:
          type: string
          enum: [operational, strategic, compliance, hr, financial, technical]
          default: operational
        owner_role:
          type: string
          description: Owner role ID
        target_date:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            field:
              type: string
...
    WorkflowListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Workflow'
        pagination:
          $ref: '#/components/schemas/Pagination'
    WorkflowResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Workflow'
        message:
          type: string
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid input
              field: content
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred